Cambio Importante: cambiar int a uint16t donde se deba

+++ include/GaloisField.h
@@ -2,7 +2,7 @@
 #define GALOIS_FIELD_H
 
 #include <vector>
-
+#include <cstdint>
 class GaloisField{
 
 private:
@@ -11,18 +11,18 @@ private:
     int size;               // 2^m
     int primitive_poly;     // Primitive Polynomial
 
-    static std::vector<int> exp_table;  // Antilog table
-    static std::vector<int> log_table;  // Log table
+    static std::vector<uint16_t> exp_table;  // Antilog table
+    static std::vector<uint16_t> log_table;  // Log table
     static int initialized_m; // Para saber si ya las calculamos
     
 public:
     GaloisField()=default;
     GaloisField(int m=0, int primitive_poly=0);
-    int add(int a, int b) const;
-    int multiply(int a, int b) const;
-    int divide(int a, int b) const;
-    int inverse(int a) const;
-    int power(int a, int e) const;
+    uint16_t add(uint16_t a, uint16_t b) const;
+    uint16_t multiply(uint16_t a, uint16_t b) const;
+    uint16_t divide(uint16_t a, uint16_t b) const;
+    uint16_t inverse(uint16_t a) const;
+    uint16_t power(uint16_t a, int e) const;
 
 
 };
